requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
about-time==4.2.1
    # via alive-progress
alembic==1.16.1
    # via optuna
alive-progress==3.2.0
    # via pymoo
autograd==1.8.0
    # via pymoo
cma==3.2.2
    # via pymoo
colorlog==6.9.0
    # via optuna
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
deap==1.4.3
    # via -r requirements.in
deprecated==1.2.18
    # via pymoo
dill==0.4.0
    # via pymoo
fonttools==4.58.0
    # via matplotlib
grapheme==0.6.0
    # via alive-progress
greenlet==3.1.1
    # via sqlalchemy
kiwisolver==1.4.7
    # via matplotlib
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via mako
matplotlib==3.10.3
    # via
    #   -r requirements.in
    #   pymoo
    #   seaborn
numpy==2.3.0
    # via
    #   -r requirements.in
    #   autograd
    #   cma
    #   contourpy
    #   deap
    #   matplotlib
    #   optuna
    #   pandas
    #   pymoo
    #   scipy
    #   seaborn
optuna==4.3.0
    # via -r requirements.in
packaging==25.0
    # via
    #   matplotlib
    #   optuna
pandas==2.3.0
    # via
    #   -r requirements.in
    #   seaborn
pillow==11.1.0
    # via matplotlib
pymoo==0.6.1.3
    # via -r requirements.in
pyparsing==3.2.3
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2024.1
    # via pandas
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   optuna
scipy==1.15.3
    # via
    #   -r requirements.in
    #   pymoo
seaborn==0.13.2
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sqlalchemy==2.0.39
    # via
    #   alembic
    #   optuna
tabulate==0.9.0
    # via -r requirements.in
tk
tqdm==4.67.1
    # via
    #   -r requirements.in
    #   optuna
typing-extensions==4.13.2
    # via
    #   alembic
    #   sqlalchemy
tzdata==2025.2
    # via pandas
wrapt==1.17.2
    # via deprecated

# The following packages are considered to be unsafe in a requirements file:
# setuptools
